// backend/src/patients/patient-status.enum.ts
export enum PatientStatus {
  NOT_SCREENED = 'not_screened', // ยังไม่ได้คัดกรอง
  SCREENED = 'screened', // คัดกรองแล้ว
  TRACKING = 'tracking', // ติดตามอยู่
  COMPLETED = 'completed', // ครบการติดตาม
}